[Admin] Modularized RowProject
